fix: zookeeper hostname resolve
ZooKeeper reads the list of servers in the zoo.cfg on startup and looks for its "own" entry.
It then uses this entry to determine which port and interface to listen on.
Since the provided hostname will resolve to 127.0.0.1 on this machine, ZooKeeper will listen on
the local loopback interface and therefore does not accept connections from the other ZooKeeper servers.